feat: Making `Split` act like a mapping
This was done in anticipation of users requiring the ability
to build a pipeline easily, in which a dictionary structure
is probably the easiest.

As a consequence, introduce a new method to
`Choice` so we can iterate over paths with
their weights. This will require further follow up
to integrate properly