feat(tasks): Limit tasks with pynisher
To allow for tasks that have their resources limited, we use `pynisher`.
This is done by sending a specially wrapped function to the `Executor`
which will then get executed there. This wrapped function spawns a new
process and executes the code in there, making sure process limits are
set.

This could potentially fail if the `Executor` is run on a different
operating system then the main process but we will not support this due
to complexity.

This also splits up the report into 3 distinct types to help typing and
the possible attributes each report would have.