feat(Step): Several convenience methods
* `path_to()` - To get a path to a step from a given step
* `find()` - Same as `Pipeline.find` but now available in `Step` natively

To enable backwards traversal for `find()`, `traverse()`, `path_to()` and the
following methods, I added a `.parent` attribute to `Step`.

Previously, a step which was part of a `Split` had no idea that it
was part of a larger DAG and so locally it thought it was the _highest_ (`head`).

When this `.parent` is set, this indicates it is the head of a path
in a `Split`. With this we can enable some more convenience methods
that wern't possible just using `.prv`.

* `climb()` - Generate an iterator of steps to reach the root of the DAG
* `root()` - Get the root of the DAG
* `qualified_name()` - Uses `path_to()` to generate a delimited name
  based on entries in the path that are splits. These coincide wih
  our currently implemented parsers.
* `traverse(backwards=)` - Enable traversel backwards

---

This commit also fixes a nasty bug that surfaced during the use of
`.parent`. Essentially use `attrs.evolve` for `attrs` classes instead
of `copy.copy`, otherwise the _post-init_ methods don't fire and
the difficulties of immutability and stale references surfaces -_-